rename websocketStore to useMultiplayerWebsocketStore
to match useSingleplayerGameStore.ts
